Payments360-Frontend/src/app/common/search-bar/search-bar.component.ts
import { Component, EventEmitter, OnInit, Output } from '@angular/core';

@Component({
  selector: 'app-search-bar',
  templateUrl: './search-bar.component.html',
  styleUrls: ['./search-bar.component.scss']
})
export class SearchBarComponent implements OnInit {

  constructor() { }

  ngOnInit(): void {
  }

  private enteredSearchValue: string = '';
  public get EnteredSearchValue(): string {
    return this.enteredSearchValue;
  }
  public set EnteredSearchValue(val:string){
    this.enteredSearchValue = val;
  };

  @Output()
  searchTextChanged: EventEmitter<string> = new EventEmitter<string>();

  onSearchTextChanged(){
    this.searchTextChanged.emit(this.enteredSearchValue);
  }
}
